+++ src/utils/evaluator.js
@@ -80,6 +80,8 @@ const PERFORM_OP_IN_BATCH_OBJ = {
 const BATCH_HIS_READ_SCHEMA = yup.object({
     ...PERFORM_OP_IN_BATCH_OBJ,
     ...getBatchProp(HIS_READ_BATCH_SIZE)
+    ...getBatchProp(HIS_READ_BATCH_SIZE),
+    ...getReturnResultProp(true)
 });
 const BATCH_HIS_WRITE_SCHEMA = yup.object({
     ...PERFORM_OP_IN_BATCH_OBJ,
@@ -203,5 +205,6 @@ module.exports = {
     CLIENT_SCHEMA,
     PERFORM_OP_IN_BATCH_SCHEMA,
     BATCH_HIS_WRITE_SCHEMA,
+    BATCH_HIS_READ_SCHEMA,
     deriveFromDefaults
 };