+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   as the basis for all new batch functions added.
 - Add new batch functions:
   - `client.batch.hisWrite(payload, options)`
+  - `client.batch.hisRead(ids, from, to, options)`
 
 ### CHANGED
 - Client no longer throws a Axios error if a response has been received and response is a Haystack of GraphQL error.